requirements.txt
catboost==1.0.3
ipython==8.4.0
joblib==1.1.0
matplotlib==3.4.3
missingno==0.4.2
networkx==2.6.3
numpy==1.20.3
optuna==2.10.0
pandas==1.3.4
plotly==5.5.0
PyYAML==6.0
scikit_learn==1.1.1
scipy==1.7.1
seaborn==0.11.2
shap==0.40.0
torch==1.11.0
tqdm==4.62.3
